Save Data Configuration
Created SavedData.cs to save the players locked/unlocked game modes as well as highScore for each. Added more variables to configure in GameMode SO, for the save data. Set up the methods for actually saving/loading/resetting the PlayerPrefs(just for gameModes and Highscores). Created a scoreKeeper, and placed UpdateScore in DetectBullets. Organized Scripts folder in project